project-LICHEN-3w7d: fix duplicate linker section registration in puck
Remove ITERABLE_SECTION_RAM(coap_resource_lichen_coap_server, ...) from
puck/sections-ram.ld. The `coap_resource_lichen_coap_server` iterable section
is owned by subsys/lichen/coap/ under CONFIG_LICHEN_COAP_SERVER_STANDALONE
and is already registered by its sections-ram.ld. Including it unconditionally
in puck/sections-ram.ld risked duplicate linker section declarations and link
failures when the subsys conditional was also active.

+++ rust/lichen-coap/src/codec.rs
@@ -727,7 +727,7 @@ mod tests {
         assert_eq!(opts[0].value, b"sensors");
         assert_eq!(opts[1].value, b"temp");
         assert_eq!(opts[2].number, OptionNumber::ContentFormat as u16);
-        assert_eq!(opts[2].as_uint().unwrap(), content_format::CBOR);
+        assert_eq!(opts[2].as_uint().unwrap(), content_format::CBOR as u32);
     }
 
     #[test]
